requirements.txt
requests>=2.31.0
python-dotenv==1.1.0

# DB — Supabase (PostgreSQL + pgvector)
supabase>=2.0.0

# RAG — 임베딩 모델
sentence-transformers>=3.0.0
torch>=2.0.0
